Separate admin and member cbank reports into distinct views.
Displaying different data based on the logged-in user is a controller
concern. Thus, the cbank controller code should determine which view
to display.

In testing this functionality, a bug in the testsuite was discovered
that allowed the population of the upstream module in test_cbank to
break other portions of the testsuite. This bug was also fixed in
this commit.


git-svn-id: https://svn.mcs.anl.gov/repos/clusterbank/trunk@496 0b55062b-bd33-0410-a79d-d5943e5cb07c