Users report is working correctly.
I learned a lot of lessons from the allocations report,
as it is likely the most complex report in cbank.
(Either that, or it ties with the projects report.)

In any case, I have applied those lessons back on the
users and projects report. All three have changed,
but at least the users report is where I want it to be.


git-svn-id: https://svn.mcs.anl.gov/repos/clusterbank/trunk@537 0b55062b-bd33-0410-a79d-d5943e5cb07c