Fixed a permission vulnerability due to variable collision.
The cbank controllers used the variable "user" to reference the current
logged-in user, but the variable "user" was also used as a placeholder
when iterating through lists of users in the same namespace. This
allowed for the unpredictable switching of the controller's perception
of who was currently logged in.

The variable "u" is now used when iterating through lists of users when
it would collide with a prior "user" variable.


git-svn-id: https://svn.mcs.anl.gov/repos/clusterbank/trunk@543 0b55062b-bd33-0410-a79d-d5943e5cb07c