Updated allocations report for #219.
The allocations report behaves exactly like the projects report. Rather,
it should, except that I found a bug in the projects report that caused
it to only show aggregate data for a the current user, always.

I've fixed it in the allocations report, and will backport the fix to
the projects report, or unify the copied code.


git-svn-id: https://svn.mcs.anl.gov/repos/clusterbank/trunk@608 0b55062b-bd33-0410-a79d-d5943e5cb07c