Tests for user_summary()
These tests are ported over from print_users_list() which used to
contain this code.  Those tests will need to be revised to test
only the formatting aspects.

Additionally, I modified user_summary() to emit charge and refund
data separately, to be combined by the view.  The cli has been
updated, but this should eventually be tested in integration tests.