Tests for project_summary()
These tests are ported over from print_projects_list() which used
to contain this code.  Those tests will need to be revised to test
only the formatting aspects.

Additionally, I modified project_summary() to emit charge and refund
data separately, to be combined by the view.  The cli has been
updated, but this should eventually be tested in integration tests.
I have not, however, split out the "available" calculation.  It might
happen in the future, but for now I'm not convinced.

I also modified project_summary() to filter holds based on allocation
availability.